Refactored rate runner, using exported interface instead of struct
