+++ pkg/f1/raterun/runner_test.go
@@ -1,4 +1,4 @@
-package raterunner
+package raterun
 
 import (
 	"testing"
